requirements.txt
numpy==1.21.0
pandas==1.3.0
gensim==4.0.1
fasttext==0.9.2
faiss-cpu==1.7.2
scikit-learn==0.24.2
matplotlib==3.4.2
seaborn==0.11.1
nltk==3.6.2
indic-nlp-library==0.4.0
wget==3.2
